ADD :: ScoreCalculator
