+++ apps/user/src/app/form/성적입력/성적입력.tsx
@@ -2,8 +2,8 @@ import {
     AttendanceCalculator,
     CertificateCalculator,
     FormController,
-    GradeCalculator,
     GradePreview,
+    ScoreCalculator,
     VolunteerCalculator,
 } from '@/components/form';
 import { SCORE_STEP_LIST } from '@/constants/form/data';
@@ -48,12 +48,12 @@ const 성적입력 = () => {
             <SwitchCase
                 value={currentScoreStep}
                 caseBy={{
-                    성적입력: <GradeCalculator />,
+                    성적입력: <ScoreCalculator />,
                     출결상황: <AttendanceCalculator />,
                     봉사시간: <VolunteerCalculator />,
                     자격증: <CertificateCalculator />,
                 }}
-                defaultComponent={<GradeCalculator />}
+                defaultComponent={<ScoreCalculator />}
             />
             <FormController
                 onPrevious={handlePreviousButtonClick}
