+++ apps/user/src/components/form/Calculators/ScoreCalculator/검정고시Calculator/검정고시CaculatorItem/검정고시CaculatorItem.tsx
