+++ apps/user/src/components/form/Calculators/index.ts
@@ -1,5 +1,4 @@
 export { default as AttendanceCalculator } from './AttendanceCalculator/AttendanceCalculator';
 export { default as CertificateCalculator } from './CertificateCalculator/CertificateCalculator';
-export { default as GradeCalculator } from './GradeCalculator/GradeCalculator';
+export { default as ScoreCalculator } from './ScoreCalculator/ScoreCalculator';
 export { default as VolunteerCalculator } from './VolunteerCalculator/VolunteerCalculator';
-export { default as 검정고시Calculator } from './검정고시Calculator/검정고시Calculator';
